+++ AdviceService/src/main/java/pl/com/mtd/adviceservice/model/Category.java
@@ -1,4 +1,4 @@
-package pl.com.mtd.AdviceService.model;
+package pl.com.mtd.adviceservice.model;
 
 import javax.persistence.*;
