+++ AdviceService/src/main/java/pl/com/mtd/adviceservice/repository/QuestionRepository.java
@@ -1,8 +1,8 @@
-package pl.com.mtd.AdviceService.repository;
+package pl.com.mtd.adviceservice.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import pl.com.mtd.AdviceService.model.Question;
+import pl.com.mtd.adviceservice.model.Question;
 
 @Repository
 public interface QuestionRepository extends JpaRepository<Question, Long> {
